added a cast to clear a compiler warning
git-svn-id: svn://orca.physics.unc.edu/Orca64@11969 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Dialogs/CommandCenter/Scripting/ORScriptRunner.m
@@ -39,7 +39,7 @@ extern uint32_t num_lines;
 extern id functionList;
 extern int OrcaScriptparse(void);
 ORScriptRunner* theScriptRunner = nil;
-int OrcaScriptYYINPUT(char* theBuffer,uint32_t maxSize) 
+int OrcaScriptYYINPUT(char* theBuffer,unsigned long maxSize) 
 {
 	return (int)[theScriptRunner yyinputToBuffer:theBuffer withSize:(int)maxSize];
 }
